Move LocDiff to lib
